Blood_test_analyser-main/backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pymongo==4.6.0
python-dotenv==1.0.0
pydantic==2.5.0
pydantic-settings==2.1.0
openai==1.3.0
pyjwt==2.8.1
python-multipart==0.0.6
passlib==1.7.4
bcrypt==4.1.1
motor==3.3.2
